requirements.txt
Flask==3.1.0
requests==2.32.3
gunicorn==23.0.0
colorama==0.4.6
python-dotenv==1.0.1
Flask-Limiter==3.8.0
Flask-Talisman==1.1.0
Flask-Caching==2.3.0
Flask-CORS==5.0.0
Flask-Compress==1.15